#27 split up extensions class, use Sort extensions in ThePageService, remove Gren Unit test
