--- ThePage/src/ThePage.UnitTests/BL/GrenUnitTest.cs
@@ -1,14 +1,0 @@
-using System;
-using Xunit;
-
-namespace ThePage.UnitTests
-{
-    public class GrenUnitTest
-    {
-        [Fact]
-        public void CheckFilledListGenreConvertedToCellGenres()
-        {
-            Assert.True(true);
-        }
-    }
-}
